move each pages compo to components folder✅
