+++ frontend/public/components/cut.js
