+++ frontend/public/components/home.js
