+++ frontend/public/template.html
@@ -31,11 +31,11 @@
         </nav>
         <!-- replace me! -->
     </main>
+    <script type='module' src='./components/home.js'></script>
+    <script type='module' src='./components/breed.js'></script>
+    <script type='module' src='./components/cut.js'></script>
     <script type='module' src='./scripts/navigation.js'></script>
-    <script type='module' src='./scripts/home.js'></script>
-    <script type='module' src='./scripts/breed.js'></script>
     <script type='module' src='./scripts/breed-manip.js'></script>
-    <script type='module' src='./scripts/cut.js'></script>
 </body>
 
 </html>
